requirements.txt
anyio==4.4.0
asgiref==3.8.1
certifi==2024.7.4
Django==4.2.15
django-debug-toolbar==4.4.6
easyocr==1.7.1
filelock==3.15.4
fsspec==2024.6.1
fuzzywuzzy==0.18.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imageio==2.35.1
Jinja2==3.1.4
lazy_loader==0.4
Levenshtein==0.25.1
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pillow==10.4.0
pyav==12.3.0
pyclipper==1.3.0.post5
python-bidi==0.6.0
python-dotenv==1.0.1
python-Levenshtein==0.25.1
python-telegram-bot==21.4
PyYAML==6.0.2
rapidfuzz==3.9.6
scikit-image==0.24.0
scipy==1.14.1
shapely==2.0.6
sniffio==1.3.1
sqlparse==0.5.1
sympy==1.13.2
tifffile==2024.8.10
torch==2.2.2
torchvision==0.17.2
typing_extensions==4.12.2
